Update video player to modal style
This commit changes the presentation of the video player from inline to a modal style. Additionally, a close button and backdrop are added for improved user control. Other minor changes include preventing the default action of the BaseVideoHolder and changes to the TechFellowsInNumbers component to extract data from hero-banner.json.